Add solution for problem `C - Boredom` Link: https://codeforces.com/contest/853/problem/C Tags: data structures, *2100 Ref: https://codeforces.com/contest/853/submission/30152796

+++ codeforces/853/C.cpp
@@ -31,7 +31,7 @@ struct BIT {
     }
 };
 
-BIT<int> bit;
+BIT<Long> bit;
 
 
 const int N = 2e5 + 5;
